requirements.txt
python-dotenv
certifi

mysql-connector-python
pymongo
pymongo[srv]

numpy
pandas

matplotlib
seaborn

scikit-learn
catboost
xgboost

kagglehub
streamlit
flask
fastapi
uvicorn
dill

pyaml

mlflow
dagshub
-e .